rules/cpp: demonstrate that depfiles are properly parsed and available in the ActionCache
As a POC, show that all required information from depfiles is available.
Use a brute-force approach to retroactively filter declaredIncludeSrcs
for the actually required include sources.

As a result, actions whose inputs haven't changed but whose declared
files have changed to now contain unused files are not executed again.

The underlying problem here is that the ActionCache does not exist or is
not used when CppCompileContexts and CppCompileActions are initialized.
These, originally mostly immutable, instances are based solely on the
view that the analysis phase gets from inspecting the project files. The
analysis phase does not take into account the information gathered - and
stored - during the previous execution phases.

This change was only tested using the following command. Other test
cases may have been broken.

  bazelisk test //src/test/shell/integration:cpp_test